requirements.txt
# ==============================================================
# FILE: titanic_agent/requirements.txt
# ==============================================================

# --------------------------------------------------------------
# Web framework
# --------------------------------------------------------------
fastapi==0.115.5
uvicorn[standard]==0.32.1

# --------------------------------------------------------------
# Data validation
# --------------------------------------------------------------
pydantic==2.9.2
pydantic-settings==2.6.1

# --------------------------------------------------------------
# Environment variables
# --------------------------------------------------------------
python-dotenv==1.0.1

# --------------------------------------------------------------
# HTTP client (dataset download + Streamlit -> FastAPI calls)
# --------------------------------------------------------------
requests==2.32.3

# --------------------------------------------------------------
# LangChain core
# --------------------------------------------------------------
langchain>=0.3.20
langchain-core>=0.3.49
langchain-community>=0.3.19

# --------------------------------------------------------------
# Groq LLM provider
# --------------------------------------------------------------
langchain-groq>=0.3.2

# --------------------------------------------------------------
# Data processing
# --------------------------------------------------------------
pandas==2.2.3
numpy==1.26.4

# --------------------------------------------------------------
# Visualisation
# --------------------------------------------------------------
matplotlib==3.9.2
seaborn==0.13.2

# --------------------------------------------------------------
# Frontend
# --------------------------------------------------------------
streamlit==1.39.0
Pillow==10.4.0
